requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
youtube-transcript-api>=0.6.2
